childcare_agency/requirements.txt
agency-swarm
python-dotenv
requests
pandas
matplotlib
nltk
pydantic 